requirements.txt
Flask
python-dotenv
langchain
langchain-openai
langchain-google-genai
langchain-deepseek
supabase
langchain-agents
pydantic